requirements.txt
flask==3.1.0
flask-login==0.6.3
paho-mqtt==1.5.1
requests==2.32.3